Crate taxonomy

Modules

edit
Functions for changing the topology or contents of a taxonomy

formats
Importers and exporters for common taxonomic formats

weights
Code for working with distance relationships in taxonomic trees and tracking weights assigned to different nodes in the tree.

Structs

GeneralTaxonomy
A concrete implementation of the Taxonomy trait suitable for interconversions of different formats.

Enums

TaxRank
A taxonomic rank. For example, a species or phylum.

TaxonomyError
A taxonomically-specific error

Traits

Taxonomy
Taxonomy is a trait exposing a number of traversal and informational methods given a smaller number of required methods. This allows different taxonomic data structures with different memory/speed requirements to have a common set of methods. T is the type for the taxonomy ID and D for the distance

Type Definitions

Result
A wrapper type for taxonomy results. Alias for Result<T, failure::Error>