Enum taxonomy::TaxRank

#[non_exhaustive]pub enum TaxRank {
    Domain,
    Subdomain,
    Realm,
    Subrealm,
    Hyperkingdom,
    Superkingdom,
    Kingdom,
    Subkingdom,
    Infrakingdom,
    Parvkingdom,
    Superphylum,
    Phylum,
    Subphylum,
    Infraphylum,
    Microphylum,
    Superclass,
    Class,
    Subclass,
    Infraclass,
    Parvclass,
    Superdivision,
    Division,
    Subdivision,
    Infradivision,
    Superlegion,
    Legion,
    Sublegion,
    Infralegion,
    Supercohort,
    Cohort,
    Subcohort,
    Infracohort,
    Superorder,
    Gigaorder,
    Magnorder,
    Grandorder,
    Mirorder,
    SeriesFish,
    Order,
    Nanorder,
    Hypoorder,
    Minorder,
    Suborder,
    Infraorder,
    Parvorder,
    Section,
    Subsection,
    Gigafamily,
    Megafamily,
    Grandfamily,
    Hyperfamily,
    Superfamily,
    Epifamily,
    SeriesLepidoptera,
    GroupLepidoptera,
    Family,
    Subfamily,
    Infrafamily,
    Supertribe,
    Tribe,
    Subtribe,
    Infratribe,
    Genus,
    Subgenus,
    Series,
    SubseriesBotany,
    SpeciesGroup,
    SpeciesSubgroup,
    Species,
    Subspecies,
    Varietas,
    Subvarietas,
    Forma,
    Subforma,
    Cultivar,
    Breed,
    Strain,
    Individual,
    Clade,
    SeroGroup,
    Biotype,
    FormaSpecialis,
    Isolate,
    Serotype,
    Genotype,
    Morph,
    Pathogroup,
    Unspecified,
}

A taxonomic rank. For example, a species or phylum.

We use this instead of a String/&str to allow stricter type-checking by forcing all taxonomic ranks to fall within the below categories (this includes all current NCBI ranks and a few others, mostly ones specific to zoology and botany).

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Domain

Subdomain

Realm

Subrealm

Hyperkingdom

Superkingdom

Kingdom

Subkingdom

Infrakingdom

Parvkingdom

Superphylum

Phylum

Subphylum

Infraphylum

Microphylum

Superclass

Class

Subclass

Infraclass

Parvclass

Superdivision

Division

Subdivision

Infradivision

Superlegion

Legion

Sublegion

Infralegion

Supercohort

Cohort

Subcohort

Infracohort

Superorder

Gigaorder

Magnorder

Grandorder

Mirorder

SeriesFish

Order

Nanorder

Hypoorder

Minorder

Suborder

Infraorder

Parvorder

Section

Subsection

Gigafamily

Megafamily

Grandfamily

Hyperfamily

Superfamily

Epifamily

SeriesLepidoptera

GroupLepidoptera

Family

Subfamily

Infrafamily

Supertribe

Tribe

Subtribe

Infratribe

Genus

Subgenus

Series

SubseriesBotany

SpeciesGroup

SpeciesSubgroup

Species

Subspecies

Varietas

Subvarietas

Forma

Subforma

Cultivar

Breed

Strain

Individual

Clade

SeroGroup

Biotype

FormaSpecialis

Isolate

Serotype

Genotype

Morph

Pathogroup

Unspecified

Implementations

impl TaxRank

pub fn to_ncbi_rank(self) -> &'static str

Coverts a TaxRank into a one of the rank strings NCBI uses. Note that this doesn't handle ranks that are not used by the NCBI taxonomy.