Function taxonomy::formats::newick::save_newick

pub fn save_newick<'t, T: 't, D: 't, X, W>(
    tax: &'t X, 
    writer: &mut W, 
    root_node: Option<T>
) -> Result<()> where
    W: Write,
    T: Clone + Debug + Display + PartialEq,
    D: Debug + Display + PartialOrd + Sum,
    X: Taxonomy<'t, T, D>, 

Write a Taxonomy object formatted as Newick to a writer.

Takes a root_node to allow writing only a subset of the total taxonomy.

Still somewhat experimental and may not support all Newick features.