Struct taxonomy::GeneralTaxonomy

pub struct GeneralTaxonomy {
    pub tax_ids: Vec<String>,
    pub parent_ids: Vec<usize>,
    pub parent_dists: Vec<f32>,
    pub ranks: Vec<TaxRank>,
    pub names: Vec<String>,
    // some fields omitted
}

A concrete implementation of the Taxonomy trait suitable for interconversions of different formats.

Fields
tax_ids: Vec<String>parent_ids: Vec<usize>parent_dists: Vec<f32>ranks: Vec<TaxRank>names: Vec<String>
Implementations

impl GeneralTaxonomy

pub fn from_arrays(
tax_ids: Vec<String>,
parent_ids: Vec<usize>,
names: Option<Vec<String>>,
ranks: Option<Vec<TaxRank>>,
dists: Option<Vec<f32>>
) -> Result<Self>

Initializer for a new GeneralTaxonomy.

All Vecs must be the same length or initialization will fail.

pub fn from_taxonomy<'t, T: 't, D: 't>(
taxonomy: &'t impl Taxonomy<'t, T, D>
) -> Result<Self> where
T: Clone + Debug + Display + PartialEq + Ord,
D: Clone + Debug + Ord + Into<f64> + PartialOrd + PartialEq + Sum,

Build a GeneralTaxonomy from any object that implements the Taxonomy trait.

pub fn from_internal_id(&self, tax_id: usize) -> Result<&str>

Given an internal tax_id (an array position) return the corresponding external tax_id (e.g. a NCBI taxonomy ID).

Because the Taxonomy trait is implemented for internal IDs also, the can be used to speed up some operations by avoiding a string lookup.

pub fn to_internal_id(&self, tax_id: &str) -> Result<usize>

Given an external tax_id (e.g. a NCBI taxonomy ID) return the corresponding internal tax_id (the position of that tax node in the internal array).

Because the Taxonomy trait is implemented for internal IDs also, the can be used to speed up some operations by avoiding a string lookup.

pub fn find_by_name(&self, name: &str) -> Result<&str>

Retrieves the external ID given its name

Always errors on Newick since it doesn't contain names.

pub fn remove(&mut self, tax_id: usize) -> Result<()>

Remove a single node from the taxonomy.

Unlike pruning the children of the node are kept, but are rejoined onto the node's parent. The root node can not be removed

pub fn add(&mut self, parent_id: usize, tax_id: &str) -> Result<()>

Add a new node to the taxonomy.