Module taxonomy::weights

Code for working with distance relationships in taxonomic trees and tracking weights assigned to different nodes in the tree.

Functions

all_weighted_paths
Calculates the summed weight of every path to root for a set of node weights given a corresponding taxonomy.

maximum_weighted_path
Find the lineage that has the greatest summed weight from all of the weights and a corresponding taxonomy.

rollup_weights
Coverts a set of weights into the set of weights including their children using the corresponding taxonomy.