Function taxonomy::weights::all_weighted_paths

pub fn all_weighted_paths<'t, D: 't, T: 't, W: 't, S: BuildHasher>(
    taxonomy: &'t impl Taxonomy<'t, T, D>, 
    weights: &HashMap<T, W, S>
) -> Result<Vec<(T, W)>> where
    D: Debug + PartialOrd + Sum,
    T: Clone + Debug + Display + Hash + PartialEq + Ord,
    W: Clone + Debug + Ord + PartialOrd + PartialEq + Sum, 

Calculates the summed weight of every path to root for a set of node weights given a corresponding taxonomy.

This is like maximum_weighted_path, but rather than just returning the best path, it returns all of the paths.