Function taxonomy::weights::maximum_weighted_path

pub fn maximum_weighted_path<'t, D: 't, T: 't, W: 't, S: BuildHasher>(
    taxonomy: &'t impl Taxonomy<'t, T, D>, 
    weights: &HashMap<T, W, S>, 
    take_first_in_tie: bool
) -> Result<Option<(T, W)>> where
    D: Debug + PartialOrd + Sum,
    T: Clone + Debug + Display + Hash + PartialEq + Ord,
    W: Clone + Debug + PartialOrd + PartialEq + Sum, 

Find the lineage that has the greatest summed weight from all of the weights and a corresponding taxonomy.

Note that this implementation doesn't use the "distances" in the taxonomy itself, but only uses the weights provided. This can greatly speed up the calculation because it reduces the number of possible paths that need to be checked.

If take_first_in_tie is set the tax_id returned will be the most specific possible (e.g. a strain), but may not be completely correct if there are multiple specific paths (e.g. multiple strains with the same path weight). Setting to false will return the common ancestor of those strains.